Register the signal handlers in Starboard
Added the signal handlers registration the same way
we had it in Cobalt 25. Also disabled the Chromium
signal handlers as they registered too late and
don't have access to the EvergreenInfo for unwinding.

Miner fixes in the symbolizer.py script.

Issue: 423906285
