+++ starboard/linux/shared/BUILD.gn
@@ -22,6 +22,7 @@ group("starboard_platform") {
     ":starboard_base_symbolize",
     ":starboard_platform_sources",
     "//starboard/common",
+    "//starboard/crashpad_wrapper:wrapper_stub",
     "//starboard/shared/ffmpeg:ffmpeg_dynamic_load",
     "//starboard/shared/starboard/media:media_util",
   ]
