+++ starboard/linux/x64x11/BUILD.gn
@@ -16,9 +16,6 @@ static_library("starboard_platform") {
   check_includes = false
 
   sources = [ "//starboard/linux/x64x11/run_starboard_main.cc" ]
-  if (!sb_is_modular || sb_is_evergreen) {
-    sources += [ "//starboard/linux/x64x11/main.cc" ]
-  }
 
   public_deps = [
     "//starboard/linux/x64x11/shared:starboard_platform",
@@ -31,7 +28,6 @@ static_library("starboard_platform") {
 if (sb_is_modular && !sb_is_evergreen) {
   static_library("starboard_platform_with_main") {
     check_includes = false
-    sources = [ "//starboard/linux/x64x11/main.cc" ]
     configs += [ "//starboard/build/config:starboard_implementation" ]
     public_deps = [ ":starboard_platform" ]
   }
